Propagate the recent changes of ubuntu2404-wsl.sh to debian12.sh

+++ debian12.sh
@@ -96,8 +96,8 @@ copy inputrc    ~/.inputrc
 
 
 # Vim, Git / Git-LFS, tree, ripgrep
-run apt-get update
-run apt-get install -y --no-install-recommends \
+run apt update
+run apt install -y --no-install-recommends \
 vim git git-lfs tree ripgrep
 
 copy vimrc.local /etc/vim/vimrc.local
@@ -106,7 +106,7 @@ run git lfs install --skip-repo
 
 
 # img2sixel
-run apt-get install -y --no-install-recommends \
+run apt install -y --no-install-recommends \
 libsixel-bin
 
 
@@ -119,16 +119,16 @@ run rye self completion '>' /usr/share/bash-completion/completions/rye
 
 
 # X window forwarding and some small programs for testing
-run apt-get install -y --no-install-recommends \
+run apt install -y --no-install-recommends \
 xauth xxd x11-apps mesa-utils
 
-if [ -z "${DISPLAY}" ]; then
-    echo -e "${COLOR_RED}\$DISPLAY is empty.  Please relogin from an SSH client to complete the process. ${COLOR_CLEAR}"
-    exit 0
+if [ -n "${DISPLAY}" ]; then
+    rm -f ~/.Xauthority
+    install --mode 0600 /dev/null ~/.Xauthority
+    run xauth add ${DISPLAY} . $(xxd -l 16 -p /dev/urandom)     # Generate ~/.Xauthority
+else
+    echo -e "${COLOR_RED}\$DISPLAY is empty... omitting to generate ~/.Xauthority${COLOR_CLEAR}"
 fi
-rm -f ~/.Xauthority
-install --mode 0600 /dev/null ~/.Xauthority
-run xauth add ${DISPLAY} . $(xxd -l 16 -p /dev/urandom)     # Generate ~/.Xauthority
 
 
 # git-delta   ref. https://github.com/dandavison/delta/releases
@@ -148,22 +148,74 @@ run systemctl start upower
 run fc-cache -fv
 
 
+# GitHub CLI
+[ -s githubcli.gpg ] ||
+run curl -o githubcli.gpg -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg
+[ -d /etc/apt/keyrings ] ||
+run install --mode 0755 --directory /etc/apt/keyrings/
+run install --mode 0644 githubcli.gpg /etc/apt/keyrings/
+
+cat > /etc/apt/sources.list.d/githubcli.list <<EOF
+deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/githubcli.gpg] \
+https://cli.github.com/packages stable main
+EOF
+
+run apt update
+run apt install -y gh
+
+
+# Claude Code
+[ -s nvm.sh ] ||
+run curl -o nvm.sh -fsSL https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh
+run bash ./nvm.sh
+
+. $HOME/.nvm/nvm.sh
+run nvm install --lts
+run nvm current
+run node -v
+run npm -v
+run npm install -g @anthropic-ai/claude-code
+run npm install -g ccusage
+
+
 # Login user settings
 #  1. Change the color of the prompt for the login user: green(32m) -> purple(35m)
 #  2. Set ~/.Xauthority
+#  3. Autoload ~/.nvm/nvm.sh
 LOGIN_USER="$(logname)"
 [ -n "$LOGIN_USER" ] || LOGIN_USER="$SUDO_USER"     # Alternative way to find the name
 if [ -n "$LOGIN_USER" ]; then
     BASHRC="~$LOGIN_USER/.bashrc"
     run [ -s $BASHRC ]
-    run sed -i -e '"/^ *PS1=/s/\[01;32m/[01;35m/"' $BASHRC
+    run sed -i $BASHRC \
+            -e '"/^ *PS1=/s/\[01;32m/[01;35m/"' \
+            -e '"/NVM_DIR/d"'
 
-    # Generate ~/.Xauthority
-    rm -f ~$LOGIN_USER/.Xauthority
-    run install --mode 0600 --owner $LOGIN_USER /dev/null ~$LOGIN_USER/.Xauthority
-    run sudo -u "$LOGIN_USER" xauth add ${DISPLAY} . $(xxd -l 16 -p /dev/urandom)
-    # Refer ~/.Xauthority from the login user
-    run echo "export XAUTHORITY=$(getent passwd "${LOGIN_USER}" | cut -d : -f 6)/.Xauthority" '>>' ~/.bashrc
+    if [ -n "${DISPLAY}" ]; then
+        # Generate ~/.Xauthority
+        rm -f ~$LOGIN_USER/.Xauthority
+        run install --mode 0600 --owner $LOGIN_USER /dev/null ~$LOGIN_USER/.Xauthority
+        run sudo -u "$LOGIN_USER" xauth add ${DISPLAY} . $(xxd -l 16 -p /dev/urandom)
+        # Refer ~/.Xauthority of the login user
+        run echo "export XAUTHORITY=$(getent passwd "${LOGIN_USER}" | cut -d : -f 6)/.Xauthority" '>>' ~/.bashrc
+    else
+        echo -e "${COLOR_RED}\$DISPLAY is empty... omitting to generate ~$LOGIN_USER/.Xauthority${COLOR_CLEAR}"
+    fi
+
+    run install --mode 0755 --owner $LOGIN_USER --directory ~$LOGIN_USER/.nvm
+    run install --mode 0644 --owner $LOGIN_USER "$HOME/.nvm/nvm.sh" ~$LOGIN_USER/.nvm/nvm.sh
+
+    # Append auto-loading of nvm.sh
+    run cat ">>" $BASHRC <<"EOF"
+export NVM_DIR="$HOME/.nvm"
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+EOF
+
+    run sudo -u $LOGIN_USER bash -i -c '"nvm install --lts"'    # nvm is a shell function.
+    run sudo -u $LOGIN_USER bash -i -c '"npm install -g @anthropic-ai/claude-code"'
+    run sudo -u $LOGIN_USER bash -i -c '"npm install -g ccusage"'
+
 else
     echo -e "${COLOR_RED}No login user found... omitting to tweak ~/.bashrc${COLOR_CLEAR}"
     echo ""
